Basic Invetory
Inventory can be interfaced with for UI by adding yourself to the
'armCountEvent' and 'legCountEvent'. Everytime the value of the arm and
leg count changes, its new value will be sent as a single integer
parameter.

Added a basic inventory system

Added a Base Limb class

Added a basic Arm class

Added a basic Leg class