Limbs Fly Off
An enemy's limbs are considered different from limb that are on the
ground, and those that are in the player's inventory since they are game
objects, but also in someone's inventory. So. They need the component
EnemyLimbBehaviour added to them. This class inherits from monobehaviour
(so it is a component) and it hold a reference to a limb object. This
means that you can give an enemy an already existant limb or create a
new one and then attach it to your enemy. When the enemy pops a limb off
of his inventory, that same limb reference is used to create a new game
object with the component LimbBehaviour at the same position and using
the same sprite as the limb the zombie had on him.

It also flies off in a random direction with a random rotation. This can
be set in 'ArmAndLeg/Assets/Settings/Limb Object Factory Settings.asset'
using the inspector.