Pick Up Limbs
You can now pick up limbs that fall from enemies. They can only be
picked up once they have 'hit the ground'. This is when their linear
velocity is less than 0.1 currently. I think a separate sprite should be
used for limbs when they are lying on the ground. The same sprite can be
used for when a zombie is moving and when the limb is flying away. Only
when it is on the ground does it need a different sprite.